src/requirements.txt
aniso8601==7.0.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-graphql-jwt==0.2.2
django-model-choices==1.0.0
graphene==2.1.8
graphene-django==2.7.1
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.8
promise==2.2.1
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
